+++ internal/provider/model_fmc_file_categories.go
@@ -21,13 +21,11 @@ package provider
 import (
 	"context"
 	"fmt"
-	"maps"
 
 	"github.com/hashicorp/go-version"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/tidwall/gjson"
-	"github.com/tidwall/sjson"
 )
 
 // End of section. //template:end imports
@@ -61,24 +59,6 @@ func (data FileCategories) getPath() string {
 // End of section. //template:end getPath
 
 // Section below is generated&owned by "gen/generator.go". //template:begin toBody
-
-func (data FileCategories) toBody(ctx context.Context, state FileCategories) string {
-	body := ""
-	if data.Id.ValueString() != "" {
-		body, _ = sjson.Set(body, "id", data.Id.ValueString())
-	}
-	if len(data.Items) > 0 {
-		body, _ = sjson.Set(body, "items", []interface{}{})
-		for key, item := range data.Items {
-			itemBody, _ := sjson.Set("{}", "name", key)
-			if !item.Id.IsNull() && !item.Id.IsUnknown() {
-				itemBody, _ = sjson.Set(itemBody, "id", item.Id.ValueString())
-			}
-			body, _ = sjson.SetRaw(body, "items.-1", itemBody)
-		}
-	}
-	return gjson.Get(body, "items").String()
-}
 
 // End of section. //template:end toBody
 
@@ -123,112 +103,16 @@ func (data *FileCategories) fromBody(ctx context.Context, res gjson.Result) {
 
 // Section below is generated&owned by "gen/generator.go". //template:begin fromBodyPartial
 
-// fromBodyPartial reads values from a gjson.Result into a tfstate model. It ignores null attributes in order to
-// uncouple the provider from the exact values that the backend API might summon to replace nulls. (Such behavior might
-// easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
-// "managed" elements, instead of all elements.
-func (data *FileCategories) fromBodyPartial(ctx context.Context, res gjson.Result) {
-	for i := range data.Items {
-		parent := &data
-		data := (*parent).Items[i]
-		parentRes := &res
-		var res gjson.Result
-
-		parentRes.Get("items").ForEach(
-			func(_, v gjson.Result) bool {
-				if v.Get("id").String() == data.Id.ValueString() && data.Id.ValueString() != "" {
-					res = v
-					return false // break ForEach
-				}
-				return true
-			},
-		)
-		if value := res.Get("id"); value.Exists() {
-			data.Id = types.StringValue(value.String())
-		} else {
-			data.Id = types.StringNull()
-		}
-		if value := res.Get("type"); value.Exists() && !data.Type.IsNull() {
-			data.Type = types.StringValue(value.String())
-		} else {
-			data.Type = types.StringNull()
-		}
-		(*parent).Items[i] = data
-	}
-}
-
 // End of section. //template:end fromBodyPartial
 
 // Section below is generated&owned by "gen/generator.go". //template:begin fromBodyUnknowns
-
-// fromBodyUnknowns updates the Unknown Computed tfstate values from a JSON.
-// Known values are not changed (usual for Computed attributes with UseStateForUnknown or with Default).
-func (data *FileCategories) fromBodyUnknowns(ctx context.Context, res gjson.Result) {
-	for i, val := range data.Items {
-		var r gjson.Result
-		res.Get("items").ForEach(
-			func(_, v gjson.Result) bool {
-				if val.Id.IsUnknown() {
-					if v.Get("name").String() == i {
-						r = v
-						return false // break ForEach
-					}
-				} else {
-					if v.Get("id").String() == val.Id.ValueString() && val.Id.ValueString() != "" {
-						r = v
-						return false // break ForEach
-					}
-				}
-
-				return true
-			},
-		)
-		if v := data.Items[i]; v.Id.IsUnknown() {
-			if value := r.Get("id"); value.Exists() {
-				v.Id = types.StringValue(value.String())
-			} else {
-				v.Id = types.StringNull()
-			}
-			data.Items[i] = v
-		}
-		if v := data.Items[i]; v.Type.IsUnknown() {
-			if value := r.Get("type"); value.Exists() {
-				v.Type = types.StringValue(value.String())
-			} else {
-				v.Type = types.StringNull()
-			}
-			data.Items[i] = v
-		}
-	}
-}
 
 // End of section. //template:end fromBodyUnknowns
 
 // Section below is generated&owned by "gen/generator.go". //template:begin Clone
 
-func (data *FileCategories) Clone() FileCategories {
-	ret := *data
-	ret.Items = maps.Clone(data.Items)
-
-	return ret
-}
-
 // End of section. //template:end Clone
 
 // Section below is generated&owned by "gen/generator.go". //template:begin toBodyNonBulk
-
-// Updates done one-by-one require different API body
-func (data FileCategories) toBodyNonBulk(ctx context.Context, state FileCategories) string {
-	// This is one-by-one update, so only one element to update is expected
-	if len(data.Items) > 1 {
-		tflog.Error(ctx, "Found more than one element to change. Only one will be changed.")
-	}
-
-	// Utilize existing toBody function
-	body := data.toBody(ctx, state)
-
-	// Get first element only
-	return gjson.Get(body, "0").String()
-}
 
 // End of section. //template:end toBodyNonBulk